Move flipper to platform modules
