--- app/src/release/kotlin/com/fuzzyfunctors/reductions/flipper/InitializeFlipper.kt
@@ -1,7 +1,0 @@
-package com.fuzzyfunctors.reductions.flipper
-
-import android.app.Application
-
-fun Application.initializeFlipper() {
-    // no op
-}
