+++ platform/flipper-api/baseline.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" ?>
+<SmellBaseline>
+  <Blacklist></Blacklist>
+  <Whitelist></Whitelist>
+</SmellBaseline>
