+++ platform/flipper-api/src/main/kotlin/com/fuzzyfunctors/reductions/flipper/api/Flipper.kt
@@ -1,0 +1,7 @@
+package com.fuzzyfunctors.reductions.flipper.api
+
+import android.app.Application
+
+interface Flipper {
+    fun initialize(application: Application)
+}
