+++ platform/flipper/src/release/kotlin/com/fuzzyfunctors/reductions/flipper/FlipperImpl.kt
@@ -1,0 +1,10 @@
+package com.fuzzyfunctors.reductions.flipper
+
+import android.app.Application
+import com.fuzzyfunctors.reductions.flipper.api.Flipper
+
+class FlipperImpl : Flipper {
+    override fun initialize(application: Application) {
+        // no-op
+    }
+}
